third_party/WebKit/LayoutTests/imported/wpt/html/webappapis/scripting/processing-model-2/window-onerror-with-cross-frame-event-listeners-1-expected.txt

CONSOLE ERROR: line 2: Uncaught ReferenceError: thereIsNoSuchCallable is not defined
This is a testharness.js-based test.
FAIL The error event from an event listener should fire on that listener's global assert_true: expected true got false
Harness: the test ran to completion.
